+++ src/options.ml
@@ -50,6 +50,7 @@ let opt_ndd   = ref false
 let opt_ne    = ref false
 let opt_sa    = ref false
 let opt_skv   = ref false
+let opt_ptc   = ref false
 let opt_raw   = ref false
 let opt_raw_whytree = ref false